+++ src/TestUtils/BlobApiTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BlobBundle\TestUtils;
 
+use Dbp\Relay\BlobBundle\DependencyInjection\DbpRelayBlobExtension;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -11,7 +12,7 @@ class BlobApiTest
 {
     public static function setUp(ContainerInterface $container): EntityManagerInterface
     {
-        return TestEntityManager::setUpEntityManager($container);
+        return TestEntityManager::setUpEntityManager($container, DbpRelayBlobExtension::ENTITY_MANAGER_ID);
     }
 
     public static function tearDown(ContainerInterface $container): void