+++ src/TestUtils/TestEntityManager.php
@@ -5,48 +5,17 @@ declare(strict_types=1);
 namespace Dbp\Relay\BlobBundle\TestUtils;
 
 use Dbp\Relay\BlobBundle\Entity\FileData;
-use Doctrine\ORM\EntityManager;
-use Doctrine\ORM\Tools\SchemaTool;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
-class TestEntityManager
+class TestEntityManager extends \Dbp\Relay\CoreBundle\TestUtils\TestEntityManager
 {
-    private EntityManager $entityManager;
-
     public function __construct(ContainerInterface $container)
     {
-        $this->entityManager = self::setUpEntityManager($container);
-    }
-
-    public static function setUpEntityManager(ContainerInterface $container): EntityManager
-    {
-        try {
-            $entityManager = $container->get('doctrine')->getManager('dbp_relay_blob_bundle');
-            assert($entityManager instanceof EntityManager);
-
-            // enable foreign key and 'ON DELETE CASCADE' support
-            $sqlStatement = $entityManager->getConnection()->prepare('PRAGMA foreign_keys = ON');
-            $sqlStatement->executeQuery();
-
-            $metaData = $entityManager->getMetadataFactory()->getAllMetadata();
-            $schemaTool = new SchemaTool($entityManager);
-            $schemaTool->updateSchema($metaData);
-        } catch (\Exception $exception) {
-            throw new \RuntimeException($exception->getMessage());
-        }
-
-        return $entityManager;
-    }
-
-    public function getEntityManager(): EntityManager
-    {
-        return $this->entityManager;
+        parent::__construct($container, 'dbp_relay_blob_bundle', FileData::class);
     }
 
     public function getFileDataById(string $identifier): ?FileData
     {
-        return $this->entityManager
-            ->getRepository(FileData::class)
-            ->find($identifier);
+        return $this->getEntityByIdentifier($identifier);
     }
 }